+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Change document upload to use the path `/v2/document` instead of `/v2/document/` (no trailing `/`).
   Both paths will continue to work in the v2 version of the API, but `/v2/document` is the intended one.
+* Made `DeepLException` and subclasses, `Usage` and `JsonFieldsStruct` constructors public, to allow for easier mocking of the `ITranslator` interface.
+  * [#40](https://github.com/DeepLcom/deepl-dotnet/issues/40) thanks to [PascalVorwerkSaixon](https://github.com/PascalVorwerkSaxion)
 
 
 ## [1.8.0] - 2023-11-03